requirements.txt
PyYAML
requests
urllib3
lxml
coverage
requests-mock
Flask>=3
werkzeug>=3
firebase-admin
mock-firestore
pytest-mock
pytest
pylint
requests-random-user-agent
jsonpath-ng
backoff
pytz
beautifulsoup4
webdriver-manager
apprise
python-dotenv
undetected-chromedriver
prompt-toolkit
ruamel.yaml
pyright
types-beautifulsoup4==4.12.0.20241020
selenium==4.27.1
gunicorn==23.0.0
setuptools
certifi
jinja2
cryptography
pydantic
exceptiongroup
